refactor(user): move shared errors down and drop the dead auth state
Two cleanups that make the code match what the thesis claims about it.

features/user imported InvalidCredentialsError and UsernameAlreadyExistsError
from features/auth in three files. Slices of one layer are supposed to stay
independent, and both errors are raised on both sides -- signing in and
changing a password, registering and editing a profile -- so they belong to
the entity the two features share. entities/user already hosts
SessionExpiredError, so the shape was there. No cross-slice import remains.

AuthState also carried a fourth case, sessionExpired, whose constructor was
never called: an expired session is turned into unauthenticated by the
listener and the user is redirected, so nothing would ever render it. The
union is now the three cases the thesis describes.
