+++ src/entities/user/index.ts
@@ -1,6 +1,8 @@
 export { AuthState } from './AuthState';
+export { InvalidCredentialsError } from './InvalidCredentialsError';
 export { SessionExpiredError } from './SessionExpiredError';
 export { authStateAtom, userAtom } from './store';
 export { User } from './User';
 export { UserId, UserIdSchema } from './UserId';
+export { UsernameAlreadyExistsError } from './UsernameAlreadyExistsError';
 export { withAuthError } from './withAuthError';
