+++ src/features/auth/index.ts
@@ -1,5 +1,4 @@
 export { AuthErrorListener } from './AuthErrorListener';
-export { InvalidCredentialsError, UsernameAlreadyExistsError } from './api/errors';
 export { LoginForm } from './LoginForm';
 export { RegisterForm } from './RegisterForm';
 export { useLogout } from './useLogout';
